Create compute linear algebra
I performed matrix multiplication on two matrices, one of which was the trend line's elements (namely the x-coefficient and y-intercept) and the other of which was the data set of Falco's frames and damage outputs. The point of multiplying the matrices out is to take Falco's inputs (frames + damage) and use linear algebra to apply a hypothesis function h(x) to them (which was generated by the linear regression trend line) to examine how effective each move is with machine learning. The trend line is a 1x2 matrix, and Falco's data is a 2x23 matrix, so the computation works out well and I received a well-defined matrix.