add test lists + variables for machine analysis
By setting up the test variables in a similar way to the tumor and non-tumor brain sets, I can plot the test training set as a scatter plot with the other variables on the other axis. That way, it makes sense visually and makes mse sum computation (which is the metric I use for linear regression from the machine) easy enough to assess the state of the brain.