add mean sum squared error, matrix + number
By taking the distance between each point and the logistic trend line, I found both the matrix component of the mean sum squared error (which was a 1D array of inputs provided by the data set) and the numerical part (which I accessed from the 2D array and for which I computed the mean). As a result, I could find a pseudo-mean for the data and add some numbers to this classification problem.