add trend line and round mean error
using sklearn.linear_model and sympy, I created a linear regression of the data. linear_model fits the two DataFrame arrays to the regression, but I had previously called slices of a regular (not DataFrame) array, so I included two extra parameters in the function to compensate. linear_model needs DataFrame and matplotlib needs regulat arrays, so I also reordered how the function is structured so the graph displays appropriate axes and has the necessary components to perform its regression. the MSE and equation also needed to be rounded to 4 significant figures, which I added to the model as well.