make changes to neural network, fix semantic error
previously the neural network ran (so no syntactic error occurred), but it generated prediction matrices with almost constant values of 0.5, which showed the network's in genuine prediction. in other words, the network showed a semantic error by not continually refreshing the forward-propagated material and just repeating the same loop. this commit uses itertools to create arrays with two randomly generated factors (such that factors a and b create arrays with [a[i], b[j]] for the neural network input) and formats the data into ordered pairs for the input and individual arrays [[x], [y], [z]] for the output. additionally, 569 values leads to a tremendous amount of itertools data to parse (about 313000 values), so I also removed 369 values (leaving 200 for the data) such that the neural network will run in PyCharm. further specific comments are left in the code.