split data into train and test sets
this commit splits the data into train and test sets. previously, I split the data for ease of use, so I made use of better neural network data-managing while doing so in order to achieve better accuracy for the model. specific commit changes are below.

1) move links to documentation.py
2) split trainSize = 369, testSize = 200
3) create train and test size for parameter1, parameter2, and output
4) change output - input_nn to:
4a) output_train - input_train for the training section with nn.train_network
and 4b) output_test - input_test for the test section
5) change from MSSE (mean sum squared error) to MSE (mean sum error), because the error was so small for the train/test split that the neural network gave me zero, and the network (unlike a linear regression or SVM) does not necessarily need to take the principal square root using MSSE