mnist cnn runs w/o error, but useless
this commit simply records that the mnist cnn model runs with train and test sets from the the loaded-in breast cancer data (though it is on debian, not on arch!), which is a huge breakthrough for this specific research part. though there is tremendous semantic error because the network is structurally and effectively useless, there is no syntactic error in this code chunk. this is particularly useful because it records that the code 1) allows for any input (jpg/png) instead of just the built-in mnist data and 2) works with the specific format for image processing that machine learning needs.

TO-DO: fix labels to correspond to binary classification (benign = 0, etc.) so the network is useful, add more iterations to increase the accuracy, and clean up/segment the code.

it is worth noting that this commit follows the linux "everything is a file" philosophy and is not actually a git file, but it does serve as a lazy yet effective way for me to store bits of data at a time because I haven't learned real GitHub version control yet :)