use shamdasani's nn format with png (CNN proxy)
I based the code off the neural network from shamdasani's tutorial and modified the input process, weights, and matrix multiplication to create a CNN proxy. The fact that I loaded arrays of pixel data from images instead of from raw data meant that the weights were useless until back-propagation took place, but once I modified the network to use the pixel data successfully it ended up running well, with no immediate signs of semantic error. I plan to leave this .git file here as a means of version control, but it is also part of the main folder in the repo as well.